Simplify u-boot boot command
The default in u-boot is to do "run distro_bootcmd", which scans devices
for boot scripts, attempts network boot, etc. We know that we always
want to launch GRUB, so this boot command does it straight away.
Also, the default boot procedure wouldn't find the boot script if the
first partition does not have the bootable flag set. While the
generated image has it set, the Debian installer might unset the
bootable flag and thus breaking the boot procedure. This commit fixes
that.